baseline
This is the first completely working nixos config I have. Nothing but the defaults.

This works as I can use firefox for all of my important work, my bookmarks and preferences won't change with any rebuilds or switches so long as I keep it (which I plan to). 

Also, nvidia drivers seem to be breaking my system, text disappears in GNOME and tabs become translucent (you can see other tabs / the desktop underneath). It was incredibly convenient to just roll back to a previous build, I'm starting to understand why people love nix so much.